Create IndexTemplate.json
This is an index template. They set settings for field mappings and data types and which ingest template to use and for what index to be applies for.

I use an anchor winlogbeat-* as you can see from the logstah pipeline. the output index it set to...

     index => "winlogbeat-%{+YYYY.MM.dd}"

As each day, month, and year increments, so does the index. Additionally, the Index pipeline is set to winlogbeat-* wild carding the part that is dynamic.